// packages/exporter/src/model/edge/decorator/arrow-type.ts
export enum ArrowType {
  // ---
  NONE = 'none',
  // --> | <--
  NOTCHED = 'notched',
  // --▷ | ◁--
  HOLLOW = 'hollow',
  // --○ | ○--
  HOLLOW_CIRCLE = 'hollow-circle',
  // --□ | □--
  HOLLOW_SQUARE = 'hollow-square',
  // --◇ | ◇--
  HOLLOW_DIAMOND = 'hollow-diamond',
  // --▶ | ◀--
  FILLED = 'filled',
  // --● | ●--
  FILLED_CIRCLE = 'filled-circle',
  // --■ | ●--
  FILLED_SQUARE = 'filled-square',
  // --◆ | ●--
  FILLED_DIAMOND = 'filled-diamond',
}